pyproject.toml
[project]
name="job"
version="0.1.0"
description=""
readme="README.md"
requires-python=">3.13"
dependencies = [
    "apify-client>=1.10.0",
    "mcp[cli]>=1.9.2",
    "openai>=1.84.0",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]